backend/requirements.txt
flask==2.3.2
flask-sqlalchemy==3.0.3
flask-migrate==4.0.4
flask-cors==3.0.10
flask-limiter==2.8.1
flask-jwt-extended==4.4.4
python-dotenv==0.21.0
psycopg2-binary==2.9.6
requests==2.28.2